requirements.txt
Django
djangorestframework
gunicorn
django-cors-headers
whitenoise
dj-database-url
psycopg2-binary
drf-yasg
